+++ docs/releases/unreleased.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+## checks
+
+- Created this module to store estimator unit testing, rather than having it in the `utils` module.
+
 ## evaluate
 
 - Refactored `evaluate.progressive_validation` to work with `base.AnomalyDetector`s.
@@ -21,3 +25,5 @@
 ## reco
 
 - Removed `reco.SurpriseWrapper` as it wasn't really useful.
+- Added a `recommend` method to each recommender.
+- Added an `is_contextual` property to each recommender.
