+++ river/_bandit/__init__.py
@@ -1,0 +1,6 @@
+from .base import Arm, Bandit, BanditPolicy
+from .epsilon_greedy import EpsilonGreedy
+from .ucb import UCB
+
+
+__all__ = ["Arm", "Bandit", "BanditPolicy", "EpsilonGreedy", "UCB"]
