+++ river/model_selection/epsilon_greedy.py
@@ -1,29 +1,9 @@
 import math
 
 from river import metrics
+from river._bandit import EpsilonGreedy
 
-from .bandit import BanditPolicy, BanditRegressor
-
-
-class EpsilonGreedy(BanditPolicy):
-    r"""$\eps$-greedy strategy."""
-
-    def __init__(self, epsilon: float, decay: float, burn_in, seed):
-        super().__init__(burn_in, seed)
-        self.epsilon = epsilon
-        self.decay = decay
-
-    def current_epsilon(self, n: int):
-        if self.decay:
-            return self.epsilon * math.exp(-n * self.decay)
-        return self.epsilon
-
-    def _pull(self, bandit):
-        yield (
-            self.rng.choice(bandit.arms)  # explore
-            if self.rng.random() < self.current_epsilon(n=bandit.n_pulls)
-            else bandit.best_arm  # exploit
-        )
+from .base import BanditRegressor
 
 
 class EpsilonGreedyRegressor(BanditRegressor):
@@ -128,7 +108,10 @@ class EpsilonGreedyRegressor(BanditRegressor):
             models=models,
             metric=metric,
             policy=EpsilonGreedy(
-                epsilon=epsilon, decay=decay, burn_in=burn_in, seed=seed,
+                epsilon=epsilon,
+                decay=decay,
+                burn_in=burn_in,
+                seed=seed,
             ),
         )
