+++ river/reco/__init__.py
@@ -3,5 +3,6 @@ from .baseline import Baseline
 from .biased_mf import BiasedMF
 from .funk_mf import FunkMF
 from .normal import RandomNormal
+from .base import Recommender
 
-__all__ = ["Baseline", "BiasedMF", "FunkMF", "RandomNormal"]
+__all__ = ["Baseline", "BiasedMF", "FunkMF", "RandomNormal", "Recommender"]
